refactor(movies-remote-data): Move out remote data logic to repository layer
